[feat] Unify eval loops Part 3/4 — EE org/events/admin cleanup and service-layer drift correction
