+++ api/ee/src/core/workspaces/types.py
@@ -3,13 +3,13 @@ from typing import Any, Dict, List, Optional
 
 from pydantic import BaseModel
 
-from ee.src.models.shared_models import Permission
+from ee.src.core.access.permissions.types import Permission
 
 
 class WorkspacePermission(BaseModel):
     # Role slugs are dynamic (env-overridable via AGENTA_ACCESS_ROLES);
     # validation against the effective scope catalog happens at the API
-    # boundary via `ee.src.core.entitlements.controls.get_role`.
+    # boundary via `ee.src.core.access.controls.get_role`.
     role_name: str
     role_description: Optional[str] = None
     permissions: Optional[List[Permission]] = None
