+++ api/ee/src/dbs/postgres/organizations/dao.py
@@ -1,9 +1,13 @@
 from typing import Optional, List
+from datetime import datetime, timezone
 
 from sqlalchemy import select, and_
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from oss.src.dbs.postgres.shared.engine import engine
+from oss.src.dbs.postgres.shared.engine import (
+    TransactionsEngine,
+    get_transactions_engine,
+)
 from ee.src.dbs.postgres.organizations.dbes import (
     OrganizationDomainDBE,
     OrganizationProviderDBE,
@@ -18,8 +22,15 @@ class OrganizationDomainsDAO:
     2. Without a session (creates own sessions): OrganizationDomainsDAO()
     """
 
-    def __init__(self, session: Optional[AsyncSession] = None):
+    def __init__(
+        self,
+        session: Optional[AsyncSession] = None,
+        engine: Optional[TransactionsEngine] = None,
+    ):
         self.session = session
+        if engine is None:
+            engine = get_transactions_engine()
+        self.engine = engine
 
     async def create(
         self,
@@ -54,7 +65,7 @@ class OrganizationDomainsDAO:
             return domain
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 domain = OrganizationDomainDBE(
                     organization_id=organization_id,
                     slug=slug,
@@ -92,7 +103,7 @@ class OrganizationDomainsDAO:
             return result.scalars().first()
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationDomainDBE).where(
                         and_(
@@ -125,7 +136,7 @@ class OrganizationDomainsDAO:
             return result.scalars().first()
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationDomainDBE).where(
                         and_(
@@ -158,7 +169,7 @@ class OrganizationDomainsDAO:
             return result.scalars().first()
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationDomainDBE).where(
                         and_(
@@ -186,7 +197,7 @@ class OrganizationDomainsDAO:
             return list(result.scalars().all())
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationDomainDBE).where(
                         OrganizationDomainDBE.organization_id == organization_id
@@ -207,9 +218,12 @@ class OrganizationDomainsDAO:
         """Update domain flags (e.g., mark as verified)."""
         domain = await self.get_by_id(domain_id=domain_id, organization_id="")
 
+        now = datetime.now(timezone.utc)
+
         if self.session:
             if domain:
                 domain.flags = flags
+                domain.updated_at = now
                 domain.updated_by_id = updated_by_id
 
                 await self.session.flush()
@@ -218,13 +232,14 @@ class OrganizationDomainsDAO:
             return domain
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 if domain:
                     # Re-attach to new session
                     domain = await session.get(OrganizationDomainDBE, domain_id)
 
                     if domain:
                         domain.flags = flags
+                        domain.updated_at = now
                         domain.updated_by_id = updated_by_id
 
                         await session.commit()
@@ -252,7 +267,7 @@ class OrganizationDomainsDAO:
             return False
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 domain = await session.get(OrganizationDomainDBE, domain_id)
 
                 if domain:
@@ -272,8 +287,15 @@ class OrganizationProvidersDAO:
     2. Without a session (creates own sessions): OrganizationProvidersDAO()
     """
 
-    def __init__(self, session: Optional[AsyncSession] = None):
+    def __init__(
+        self,
+        session: Optional[AsyncSession] = None,
+        engine: Optional[TransactionsEngine] = None,
+    ):
         self.session = session
+        if engine is None:
+            engine = get_transactions_engine()
+        self.engine = engine
 
     async def create(
         self,
@@ -311,7 +333,7 @@ class OrganizationProvidersDAO:
             return provider
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 provider = OrganizationProviderDBE(
                     organization_id=organization_id,
                     slug=slug,
@@ -350,7 +372,7 @@ class OrganizationProvidersDAO:
             return result.scalars().first()
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationProviderDBE).where(
                         and_(
@@ -378,7 +400,7 @@ class OrganizationProvidersDAO:
             return result.scalars().first()
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationProviderDBE).where(
                         OrganizationProviderDBE.id == provider_id
@@ -408,7 +430,7 @@ class OrganizationProvidersDAO:
             return result.scalars().first()
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationProviderDBE).where(
                         and_(
@@ -436,7 +458,7 @@ class OrganizationProvidersDAO:
             return list(result.scalars().all())
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 result = await session.execute(
                     select(OrganizationProviderDBE).where(
                         OrganizationProviderDBE.organization_id == organization_id
@@ -458,6 +480,8 @@ class OrganizationProvidersDAO:
         #
     ) -> Optional[OrganizationProviderDBE]:
         """Update a provider's secret reference or flags."""
+        now = datetime.now(timezone.utc)
+
         if self.session:
             provider = await self.session.get(OrganizationProviderDBE, provider_id)
 
@@ -466,6 +490,7 @@ class OrganizationProvidersDAO:
                     provider.secret_id = secret_id
                 if flags is not None:
                     provider.flags = flags
+                provider.updated_at = now
                 if updated_by_id:
                     provider.updated_by_id = updated_by_id
 
@@ -475,7 +500,7 @@ class OrganizationProvidersDAO:
             return provider
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 provider = await session.get(OrganizationProviderDBE, provider_id)
 
                 if provider:
@@ -483,6 +508,7 @@ class OrganizationProvidersDAO:
                         provider.secret_id = secret_id
                     if flags is not None:
                         provider.flags = flags
+                    provider.updated_at = now
                     if updated_by_id:
                         provider.updated_by_id = updated_by_id
 
@@ -511,7 +537,7 @@ class OrganizationProvidersDAO:
             return False
 
         else:
-            async with engine.core_session() as session:
+            async with self.engine.session() as session:
                 provider = await session.get(OrganizationProviderDBE, provider_id)
 
                 if provider:
