+++ api/ee/src/services/admin_manager.py
@@ -7,7 +7,7 @@ from sqlalchemy.future import select
 
 from oss.src.utils.logging import get_module_logger
 
-from oss.src.dbs.postgres.shared.engine import engine
+from oss.src.dbs.postgres.shared.engine import get_transactions_engine
 
 from oss.src.models.db_models import UserDB
 from oss.src.services.api_key_service import create_api_key
@@ -88,7 +88,7 @@ class ProjectRequest(BaseModel):
 
 # Role slugs are dynamic at runtime (env-overridable via AGENTA_ACCESS_ROLES);
 # validation against the effective per-scope catalog happens at the handler
-# boundary via `ee.src.core.entitlements.controls.get_role`.
+# boundary via `ee.src.core.access.controls.get_role`.
 class OrganizationMembershipRequest(BaseModel):
     role: str
     is_demo: bool
@@ -119,7 +119,9 @@ Credentials = str
 async def check_user(
     request: UserRequest,
 ) -> Optional[UserRequest]:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(UserDB).filter_by(
                 email=request.email,
@@ -136,7 +138,9 @@ async def check_user(
 async def create_user(
     request: UserRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         user_db = UserDB(
             # id=uuid7()  # use default
             #
@@ -162,7 +166,9 @@ async def create_user(
 async def create_organization(
     request: OrganizationRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         organization_db = OrganizationDB(
             name=request.name,
             description=request.description,
@@ -190,7 +196,9 @@ async def create_organization(
 async def create_workspace(
     request: WorkspaceRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         workspace_db = WorkspaceDB(
             # id=uuid7()  # use default
             #
@@ -219,7 +227,9 @@ async def create_workspace(
 async def create_project(
     request: ProjectRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         project_db = ProjectDB(
             # id=uuid7()  # use default
             #
@@ -250,7 +260,9 @@ async def create_project(
 async def create_organization_membership(
     request: OrganizationMembershipRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         membership_db = OrganizationMembershipDB(
             # id=uuid7()  # use default
             #
@@ -293,7 +305,9 @@ async def create_organization_membership(
 async def create_workspace_membership(
     request: WorkspaceMembershipRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         workspace = await session.execute(
             select(WorkspaceDB).filter_by(
                 id=request.workspace_ref.id,
@@ -334,7 +348,9 @@ async def create_workspace_membership(
 async def create_project_membership(
     request: ProjectMembershipRequest,
 ) -> Reference:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         project = await session.execute(
             select(ProjectDB).filter_by(
                 id=request.project_ref.id,
