+++ api/ee/src/services/db_manager_ee.py
@@ -1,8 +1,7 @@
-from typing import List, Set, Union, NoReturn, Optional, Tuple
+from typing import Any, Dict, List, Set, Union, NoReturn, Optional, Tuple
 import uuid
 from datetime import datetime, timezone
 
-import sendgrid
 from fastapi import HTTPException
 
 from sqlalchemy import delete, func, update
@@ -14,7 +13,9 @@ from sqlalchemy.exc import IntegrityError
 
 from oss.src.utils.logging import get_module_logger
 
-from oss.src.dbs.postgres.shared.engine import engine
+from oss.src.dbs.postgres.shared.engine import (
+    get_transactions_engine,
+)
 from oss.src.services import db_manager
 from ee.src.core.workspaces.types import (
     UserRole,
@@ -27,8 +28,12 @@ from ee.src.core.organizations.types import (
     CreateOrganization,
     OrganizationUpdate,
 )
-from ee.src.models.shared_models import WorkspaceRole
-from ee.src.core.entitlements.controls import get_roles
+from ee.src.core.access.permissions.types import Permission, RequiredRole
+from ee.src.core.access.controls import (
+    get_roles,
+    get_role_description,
+    get_role_permissions,
+)
 
 from oss.src.models.db_models import (
     OrganizationDB,
@@ -44,6 +49,7 @@ from ee.src.models.db_models import (
 from oss.src.models.db_models import (
     UserDB,
     InvitationDB,
+    DeploymentDB,
 )
 
 from ee.src.core.organizations.exceptions import (
@@ -54,14 +60,9 @@ from ee.src.dbs.postgres.organizations.dao import (
     OrganizationProvidersDAO,
     OrganizationDomainsDAO,
 )
-from ee.src.services.converters import get_workspace_in_format
-from ee.src.services.selectors import get_org_default_workspace
 
 from oss.src.utils.env import env
 
-
-# Initialize sendgrid api client
-sg = sendgrid.SendGridAPIClient(api_key=env.sendgrid.api_key)
 
 log = get_module_logger(__name__)
 
@@ -77,7 +78,9 @@ async def get_organization(organization_id: str) -> OrganizationDB:
         OrganizationDB: The fetched organization.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(OrganizationDB).filter_by(id=uuid.UUID(organization_id))
         )
@@ -96,7 +99,9 @@ async def get_organizations_by_list_ids(organization_ids: List) -> List[Organiza
         List: A list of dictionaries representing the retrieved organizations.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         organization_uuids = [
             uuid.UUID(organization_id) for organization_id in organization_ids
         ]
@@ -116,7 +121,9 @@ async def count_organizations_by_owner(owner_id: str) -> int:
     Returns:
         int: The count of organizations owned by the user.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(func.count(OrganizationDB.id)).where(
                 OrganizationDB.owner_id == uuid.UUID(owner_id)
@@ -136,7 +143,9 @@ async def get_default_workspace_id(user_id: str) -> str:
         str: The default workspace ID.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(WorkspaceMemberDB)
             .filter_by(user_id=uuid.UUID(user_id))
@@ -157,7 +166,7 @@ async def get_default_workspace_id(user_id: str) -> str:
             (
                 membership
                 for membership in memberships
-                if membership.role == WorkspaceRole.OWNER
+                if membership.role == RequiredRole.OWNER
             ),
             None,
         )
@@ -181,7 +190,9 @@ async def get_organization_workspaces(organization_id: str):
         organization_id (str): The ID of the organization
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(WorkspaceDB)
             .filter_by(organization_id=uuid.UUID(organization_id))
@@ -199,7 +210,9 @@ async def get_workspace_members(workspace_id: str) -> List[WorkspaceMemberDB]:
 
     Used by RBAC / admin helpers to derive roles and permissions.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(WorkspaceMemberDB).where(
                 WorkspaceMemberDB.workspace_id == workspace_id
@@ -221,7 +234,7 @@ async def get_workspace_administrators(workspace: WorkspaceDB) -> List[UserDB]:
     admin_user_ids = [
         str(member.user_id)
         for member in members
-        if member.role in (WorkspaceRole.ADMIN, WorkspaceRole.OWNER)
+        if member.role in (RequiredRole.ADMIN, RequiredRole.OWNER)
     ]
 
     administrators: List[UserDB] = []
@@ -385,7 +398,8 @@ async def sync_workspace_members_to_project(
         await _sync(session)
         return
 
-    async with engine.core_session() as new_session:
+    engine = get_transactions_engine()
+    async with engine.session() as new_session:
         await _sync(new_session)
 
 
@@ -402,7 +416,9 @@ async def get_default_workspace_id_from_organization(
         str: The default (first) workspace ID.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         workspace_query = await session.execute(
             select(WorkspaceDB)
             .where(
@@ -433,7 +449,9 @@ async def get_project_by_workspace(
     """
 
     assert workspace_id is not None, "Workspace ID is required to retrieve project"
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         stmt = select(ProjectDB).where(
             ProjectDB.workspace_id == uuid.UUID(workspace_id),
         )
@@ -494,7 +512,9 @@ async def create_project_member(
 async def fetch_project_memberships_by_user_id(
     user_id: str,
 ) -> List[ProjectMemberDB]:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(ProjectMemberDB)
             .filter_by(user_id=uuid.UUID(user_id))
@@ -622,7 +642,9 @@ async def create_workspace(
         user = await db_manager.get_user(user_uid)
         organization = await get_organization(organization_id)
 
-        async with engine.core_session() as session:
+        engine = get_transactions_engine()
+
+        async with engine.session() as session:
             user_result = await session.execute(select(UserDB).filter_by(uid=user_uid))
             user = user_result.scalars().first()
 
@@ -652,7 +674,9 @@ async def update_workspace(
         payload (UpdateWorkspace): The data to update the workspace with.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(select(WorkspaceDB).filter_by(id=workspace.id))
         workspace = result.scalars().first()
 
@@ -681,7 +705,9 @@ async def check_user_in_workspace_with_email(email: str, workspace_id: str) -> b
         Exception: If there is an error checking if the user belongs to the workspace.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(WorkspaceMemberDB)
             .join(UserDB, UserDB.id == WorkspaceMemberDB.user_id)
@@ -721,7 +747,9 @@ async def update_user_roles(
             f"No projects found for the provided workspace_id {workspace_id}"
         )
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         workspace_member_result = await session.execute(
             select(WorkspaceMemberDB).filter_by(
                 workspace_id=uuid.UUID(workspace_id), user_id=user.id
@@ -804,7 +832,9 @@ async def add_user_to_workspace_and_org(
     if project and str(project.workspace_id) != str(workspace.id):
         raise ValueError("Project does not belong to the provided workspace")
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         # create joined organization for user
         user_organization = OrganizationMemberDB(
             user_id=user.id, organization_id=organization.id
@@ -910,7 +940,9 @@ async def remove_user_from_workspace(
         )
     project_ids = [project.id for project in projects]
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         if not user:  # User is an invited user who has not yet created an account and therefore does not have a user object
             pass
         else:
@@ -1048,7 +1080,9 @@ async def create_organization(
         Exception: If there is an error creating the organization.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         create_org_data = payload.model_dump(exclude_unset=True)
 
         is_demo = create_org_data.pop("is_demo", False)
@@ -1137,7 +1171,9 @@ async def update_organization(
         Exception: If there is an error updating the organization.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(OrganizationDB).filter_by(id=uuid.UUID(organization_id))
         )
@@ -1299,7 +1335,9 @@ async def delete_organization(organization_id: str) -> bool:
     Raises:
         NoResultFound: If organization not found.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(OrganizationDB).filter_by(id=uuid.UUID(organization_id))
         )
@@ -1324,7 +1362,9 @@ async def delete_invitation(invitation_id: str) -> bool:
         bool: True if the invitation was successfully deleted, False otherwise.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(InvitationDB).filter_by(id=uuid.UUID(invitation_id))
         )
@@ -1386,7 +1426,9 @@ async def mark_invitation_as_used(
         HTTPException: If there is an error marking the invitation as used.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(InvitationDB).filter_by(
                 project_id=uuid.UUID(project_id), token=invitation.token
@@ -1477,7 +1519,9 @@ async def get_project_invitations(project_id: str, **kwargs):
         project_id (str): The ID of the project
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         stmt = select(InvitationDB).filter(
             InvitationDB.project_id == uuid.UUID(project_id)
         )
@@ -1497,7 +1541,9 @@ async def get_all_pending_invitations(email: str):
         email (str): The email address of the user.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(InvitationDB).filter(
                 InvitationDB.email == email,
@@ -1522,7 +1568,9 @@ async def get_project_invitation(
         InvitationDB: invitation object
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(InvitationDB).filter_by(
                 project_id=uuid.UUID(project_id), token=token, email=email
@@ -1539,7 +1587,9 @@ async def get_project_members(project_id: str):
         project_id (str): The ID of the project
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         members_query = await session.execute(
             select(ProjectMemberDB)
             .filter(ProjectMemberDB.project_id == uuid.UUID(project_id))
@@ -1567,7 +1617,9 @@ async def project_member_exists(
         True if the user belongs to the project, False otherwise.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         stmt = select(
             select(ProjectMemberDB.id)
             .filter(
@@ -1598,7 +1650,9 @@ async def workspace_member_exists(
         True if the user belongs to the workspace, False otherwise.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         stmt = select(
             select(WorkspaceMemberDB.id)
             .filter(
@@ -1645,7 +1699,9 @@ async def create_org_workspace_invitation(
     if not project:
         raise Exception(f"No project found with ID {project_id}")
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         invitation = InvitationDB(
             token=token,
             email=email,
@@ -1686,7 +1742,9 @@ async def add_user_to_organization(
     role: str = "viewer",
     # is_demo: bool = False,
 ) -> None:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         organization_member = OrganizationMemberDB(
             user_id=user_id,
             organization_id=organization_id,
@@ -1712,7 +1770,9 @@ async def add_user_to_workspace(
     role: str,
     # is_demo: bool = False,
 ) -> None:
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         # fetch workspace by workspace_id (SQL)
         stmt = select(WorkspaceDB).filter_by(id=workspace_id)
         workspace = await session.execute(stmt)
@@ -1753,7 +1813,9 @@ async def add_user_to_project(
     if not project:
         raise Exception(f"No project found with ID {project_id}")
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         project_member = ProjectMemberDB(
             user_id=user_id,
             project_id=project_id,
@@ -1793,7 +1855,9 @@ async def transfer_organization_ownership(
     Raises:
         ValueError: If new owner is not a member of the organization
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         # Verify organization exists
         org_result = await session.execute(
             select(OrganizationDB).filter_by(id=uuid.UUID(organization_id))
@@ -1925,7 +1989,9 @@ async def transfer_organization_ownership(
 
 async def admin_delete_org_membership(membership_id: uuid.UUID) -> bool:
     """Delete an org membership by ID. Returns False if not found."""
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(OrganizationMemberDB).filter_by(id=membership_id)
         )
@@ -1939,7 +2005,9 @@ async def admin_delete_org_membership(membership_id: uuid.UUID) -> bool:
 
 async def admin_delete_workspace_membership(membership_id: uuid.UUID) -> bool:
     """Delete a workspace membership by ID. Returns False if not found."""
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(WorkspaceMemberDB).filter_by(id=membership_id)
         )
@@ -1953,7 +2021,9 @@ async def admin_delete_workspace_membership(membership_id: uuid.UUID) -> bool:
 
 async def admin_delete_project_membership(membership_id: uuid.UUID) -> bool:
     """Delete a project membership by ID. Returns False if not found."""
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(ProjectMemberDB).filter_by(id=membership_id)
         )
@@ -1970,7 +2040,9 @@ async def admin_get_member_org_ids(
     org_ids: List[uuid.UUID],
 ) -> Set[uuid.UUID]:
     """Return the subset of org_ids where the user has a membership row."""
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         rows = (
             (
                 await session.execute(
@@ -1997,7 +2069,9 @@ async def admin_swap_org_memberships(
     a membership row.  For each qualifying org, target gets source's role and
     source gets target's prior role.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         source_rows = (
             (
                 await session.execute(
@@ -2059,7 +2133,9 @@ async def admin_swap_workspace_memberships(
     have a membership row.  For each qualifying workspace, target gets source's
     role and source gets target's prior role.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         source_rows = (
             (
                 await session.execute(
@@ -2121,7 +2197,9 @@ async def admin_swap_project_memberships(
     have a membership row.  For each qualifying project, target gets source's
     role and source gets target's prior role.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         source_rows = (
             (
                 await session.execute(
@@ -2177,7 +2255,9 @@ async def admin_delete_user_memberships(user_id: uuid.UUID) -> None:
 
     Called before hard-deleting a user so FK constraints are not violated.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         await session.execute(
             delete(OrganizationMemberDB).where(OrganizationMemberDB.user_id == user_id)
         )
@@ -2188,3 +2268,258 @@ async def admin_delete_user_memberships(user_id: uuid.UUID) -> None:
             delete(ProjectMemberDB).where(ProjectMemberDB.user_id == user_id)
         )
         await session.commit()
+
+
+# Merged from ee/src/services/selectors.py
+async def get_user_org_and_workspace_id(user_uid) -> Dict[str, Union[str, List[str]]]:
+    """
+    Retrieves the user ID and organization IDs associated with a given user UID.
+
+    Args:
+        user_uid (str): The UID of the user.
+
+    Returns:
+        dict: A dictionary containing the user UID, ID, list of workspace IDS and list of organization IDS associated with a user.
+              If the user is not found, returns None
+
+    Example Usage:
+        result = await get_user_org_and_workspace_id("user123")
+
+    Output:
+        { "id": "123", "uid": "user123", "organization_ids": [], "workspace_ids": []}
+    """
+
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
+        user = await db_manager.get_user_with_id(user_id=user_uid)
+        if not user:
+            raise NoResultFound(f"User with uid {user_uid} not found")
+
+        user_org_result = await session.execute(
+            select(OrganizationMemberDB)
+            .filter_by(user_id=user.id)
+            .options(load_only(OrganizationMemberDB.organization_id))  # type: ignore
+        )
+        orgs = user_org_result.scalars().all()
+        organization_ids = [str(user_org.organization_id) for user_org in orgs]
+
+        member_in_workspaces_result = await session.execute(
+            select(WorkspaceMemberDB)
+            .filter_by(user_id=user.id)
+            .options(load_only(WorkspaceMemberDB.workspace_id))  # type: ignore
+        )
+        workspaces_ids = [
+            str(user_workspace.workspace_id)
+            for user_workspace in member_in_workspaces_result.scalars().all()
+        ]
+
+        return {
+            "id": str(user.id),
+            "uid": str(user.uid),
+            "workspace_ids": workspaces_ids,
+            "organization_ids": organization_ids,
+        }
+
+
+async def user_exists(user_email: str) -> bool:
+    """Check if user exists in the database.
+
+    Arguments:
+        user_email (str): The email address of the logged-in user
+
+    Returns:
+        bool: confirming if the user exists or not.
+    """
+
+    user = await db_manager.get_user_with_email(email=user_email)
+    return False if not user else True
+
+
+async def get_org_default_workspace(organization: Organization) -> WorkspaceDB:
+    """Get's the default workspace for an organization from the database.
+
+    Arguments:
+        organization (Organization): The organization
+
+    Returns:
+        WorkspaceDB: Instance of WorkspaceDB
+    """
+
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
+        result = await session.execute(
+            select(WorkspaceDB).filter_by(
+                organization_id=organization.id,
+                type="default",
+            )
+        )
+        workspace = result.scalars().first()
+        if workspace is not None:
+            return workspace
+
+        result = await session.execute(
+            select(WorkspaceDB).filter_by(
+                organization_id=organization.id,
+            )
+        )
+        return result.scalars().first()
+
+
+# Merged from ee/src/services/db_manager.py
+async def create_deployment(
+    app_id: str,
+    project_id: str,
+    uri: str,
+) -> DeploymentDB:
+    """Create a new deployment.
+    Args:
+        app_id (str): The app variant to create the deployment for.
+        project_id (str): The project variant to create the deployment for.
+        uri (str): The URI of the service.
+    Returns:
+        DeploymentDB: The created deployment.
+    """
+
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
+        try:
+            deployment = DeploymentDB(
+                app_id=uuid.UUID(app_id),
+                project_id=uuid.UUID(project_id),
+                uri=uri,
+            )
+
+            session.add(deployment)
+            await session.commit()
+            await session.refresh(deployment)
+
+            return deployment
+        except Exception as e:
+            raise Exception(f"Error while creating deployment: {e}")
+
+
+# Merged from ee/src/services/converters.py
+def _role_slug(role: Any) -> str:
+    """Normalize an enum or string role to its slug form."""
+    return role.value if hasattr(role, "value") else str(role)
+
+
+def _expand_permissions(slugs: List[str]) -> List[str]:
+    """Expand the `"*"` wildcard to the full list of Permission enum values.
+
+    Why: `WorkspacePermission.permissions` is typed as `List[Permission]` and
+    the owner role stores `["*"]` as a wildcard. Pydantic rejects `"*"` since
+    it's not an enum member, so we materialize it at the API boundary.
+    """
+    if "*" not in slugs:
+        return slugs
+    return [p.value for p in Permission]
+
+
+async def get_workspace_in_format(
+    workspace: WorkspaceDB,
+    include_members: bool = True,
+) -> WorkspaceResponse:
+    """Converts the workspace object to the WorkspaceResponse model.
+
+    Arguments:
+        workspace (WorkspaceDB): The workspace object
+        include_members (bool): Whether to include workspace members. Defaults to True.
+
+    Returns:
+        WorkspaceResponse: The workspace object in the WorkspaceResponse model
+    """
+
+    members = []
+
+    if include_members:
+        project = await get_project_by_workspace(workspace_id=str(workspace.id))
+        project_members = await get_project_members(project_id=str(project.id))
+        invitations = await get_project_invitations(
+            project_id=str(project.id), invitation_used=False
+        )
+
+        if len(invitations) > 0:
+            for invitation in invitations:
+                if not invitation.used and str(invitation.project_id) == str(
+                    project.id
+                ):
+                    user = await db_manager.get_user_with_email(invitation.email)
+                    member_dict = {
+                        "user": {
+                            "id": str(user.id) if user else invitation.email,
+                            "email": user.email if user else invitation.email,
+                            "username": (
+                                user.username
+                                if user
+                                else invitation.email.split("@")[0]
+                            ),
+                            "status": (
+                                "pending"
+                                if invitation.expiration_date
+                                > datetime.now(timezone.utc)
+                                else "expired"
+                            ),
+                            "created_at": (
+                                str(user.created_at)
+                                if user
+                                else (
+                                    str(invitation.created_at)
+                                    if str(invitation.created_at)
+                                    else None
+                                )
+                            ),
+                        },
+                        "roles": [
+                            {
+                                "role_name": invitation.role,
+                                "role_description": get_role_description(
+                                    "workspace", _role_slug(invitation.role)
+                                ),
+                            }
+                        ],
+                    }
+                    members.append(member_dict)
+
+        for project_member in project_members:
+            member_role = project_member.role
+            member_dict = {
+                "user": {
+                    "id": str(project_member.user.id),
+                    "email": project_member.user.email,
+                    "username": project_member.user.username,
+                    "status": "member",
+                    "created_at": str(project_member.user.created_at),
+                },
+                "roles": (
+                    [
+                        {
+                            "role_name": member_role,
+                            "role_description": get_role_description(
+                                "project", _role_slug(member_role)
+                            ),
+                            "permissions": _expand_permissions(
+                                get_role_permissions("project", _role_slug(member_role))
+                            ),
+                        }
+                    ]
+                    if member_role
+                    else []
+                ),
+            }
+            members.append(member_dict)
+
+    workspace_response = WorkspaceResponse(
+        id=str(workspace.id),
+        name=workspace.name,
+        description=workspace.description,
+        type=workspace.type,
+        members=members,
+        organization=str(workspace.organization_id),
+        created_at=str(workspace.created_at),
+        updated_at=str(workspace.updated_at),
+    )
+    return workspace_response
