+++ api/oss/src/core/events/streaming.py
@@ -4,7 +4,6 @@ from uuid import UUID
 
 from orjson import dumps, loads
 from pydantic import BaseModel
-from redis.asyncio import Redis
 
 try:
     from asyncpg.pgproto.pgproto import UUID as AsyncpgUUID
@@ -12,7 +11,7 @@ except ImportError:
     AsyncpgUUID = None
 
 from oss.src.core.events.dtos import Event
-from oss.src.utils.env import env
+from oss.src.dbs.redis.shared.engine import get_streams_engine
 from oss.src.utils.logging import get_module_logger
 
 log = get_module_logger(__name__)
@@ -24,16 +23,9 @@ def _orjson_default(obj):
     raise TypeError(f"Type is not JSON serializable: {type(obj)}")
 
 
-_redis: Optional[Redis] = None
-
-
-def _get_redis() -> Optional[Redis]:
-    global _redis
-
-    if _redis is None and env.redis.uri_durable:
-        _redis = Redis.from_url(env.redis.uri_durable, decode_responses=False)
-
-    return _redis
+def _get_redis():
+    engine = get_streams_engine()
+    return engine.get_redis() if engine else None
 
 
 class EventMessage(BaseModel):
