+++ api/oss/src/routers/api_key_router.py
@@ -7,8 +7,8 @@ from oss.src.services import api_key_service
 from oss.src.models.api.api_models import ListAPIKeysResponse
 
 if is_ee():
-    from ee.src.models.shared_models import Permission
-    from ee.src.utils.permissions import check_action_access
+    from ee.src.core.access.permissions.types import Permission
+    from ee.src.core.access.permissions.service import check_action_access
 
 
 router = APIRouter()
