+++ api/oss/src/services/api_key_service.py
@@ -13,9 +13,7 @@ from sqlalchemy.orm import joinedload
 
 from oss.src.utils.logging import get_module_logger
 from oss.src.models.db_models import APIKeyDB, UserDB
-from oss.src.dbs.postgres.shared.engine import engine
-
-# from oss.src.utils.redis_utils import redis_connection
+from oss.src.dbs.postgres.shared.engine import get_transactions_engine
 
 log = get_module_logger(__name__)
 
@@ -38,7 +36,9 @@ async def _generate_unique_prefix():
     # Define the characters to use for the prefix
     alphabet = string.ascii_letters + string.digits
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         while True:
             # Generate a random 8-character prefix
             prefix = "".join(secrets.choice(alphabet) for _ in range(8))
@@ -82,7 +82,9 @@ async def create_api_key(
     # get rate limit from env
     rate_limit = 0
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         # Create an APIKeyDB instance with the prefix, hashed API key, and user_id
         api_key = APIKeyDB(
             prefix=prefix,
@@ -115,7 +117,9 @@ async def is_valid_api_key(key: str):
     - The API Key object if the API key is valid, False otherwise.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         # Check if the API key is valid (not blacklisted and not expired)
         result = await session.execute(
             select(APIKeyDB)
@@ -134,37 +138,6 @@ async def is_valid_api_key(key: str):
             return False
 
         return api_key
-
-
-# async def check_rate_limit(api_key_obj: APIKeyDB, cache_key: str):
-#     """
-#     Checks if an API key has exceeded its rate limit.
-
-#     Args:
-#     - key: The API key to be checked.
-
-#     Returns:
-#     - True if the API key has exceeded its rate limit, False otherwise.
-#     """
-
-#     if api_key_obj.rate_limit > 0:
-#         # Check rate limiting in Redis
-#         r = redis_connection()
-#         if r is not None:
-#             api_requests_within_minute = r.get(cache_key)
-#             if api_requests_within_minute is None:
-#                 # Initialize the count in Redis with an initial value of 1 and a one-minute TTL
-#                 r.setex(cache_key, 60, 1)
-#             else:
-#                 # Check if requests made within the last minute exceed the rate limit
-#                 count_within_minute = int(api_requests_within_minute.decode("utf-8"))
-#                 if count_within_minute > api_key_obj.rate_limit:
-#                     return True
-
-#         # increment the apikey usage count in redis
-#         r.incr(cache_key)
-
-#     return False
 
 
 async def use_api_key(key: str) -> Union[APIKeyDB, bool]:
@@ -234,7 +207,9 @@ async def list_api_keys(user_id: str, project_id: str) -> List[APIKeyDB]:
         List[APIKeyDB]: A list of APIKeyDB objects associated with the user, sorted by most recently created first.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(APIKeyDB)
             .filter_by(
@@ -260,7 +235,9 @@ async def delete_api_key(user_id: str, key_prefix: str):
         KeyError: If the API key does not exist or does not belong to the user.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(
             select(APIKeyDB).filter_by(
                 created_by_id=uuid.UUID(user_id), prefix=key_prefix
