--- api/oss/src/services/templates/send_email.html
@@ -1,7 +1,0 @@
-<p>Hello,</p>
-<p>
-  {username_placeholder} has {action_placeholder} {workspace_placeholder} on
-  Agenta.
-</p>
-<p>{call_to_action}</p>
-<p>Thank you for using Agenta!</p>
