+++ api/oss/src/services/user_service.py
@@ -5,9 +5,10 @@ from supertokens_python.recipe.emailpassword.asyncio import create_reset_passwor
 from oss.src.utils.env import env
 from oss.src.models.db_models import UserDB
 from oss.src.utils.logging import get_module_logger
-from oss.src.dbs.postgres.shared.engine import engine
+from oss.src.dbs.postgres.shared.engine import get_transactions_engine
 from oss.src.models.api.user_models import UserUpdate
-from oss.src.services import db_manager, email_service
+from oss.src.services import db_manager
+from oss.src.utils import emailing
 
 log = get_module_logger(__name__)
 
@@ -36,7 +37,9 @@ async def delete_user(user_id: str) -> None:
     Raises:
         NoResultFound: If user with the given ID is not found.
     """
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(select(UserDB).filter_by(id=user_id))
         user = result.scalars().first()
 
@@ -68,7 +71,9 @@ async def create_new_user(payload: dict) -> UserDB:
 
     # Attempt to create new user
     try:
-        async with engine.core_session() as session:
+        engine = get_transactions_engine()
+
+        async with engine.session() as session:
             user = UserDB(**payload)
 
             session.add(user)
@@ -110,7 +115,9 @@ async def update_user(user_uid: str, payload: UserUpdate) -> UserDB:
         NoResultFound: User with session id xxxx not found.
     """
 
-    async with engine.core_session() as session:
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
         result = await session.execute(select(UserDB).filter_by(uid=user_uid))
         user = result.scalars().first()
 
@@ -151,20 +158,11 @@ async def generate_user_password_reset_link(user_id: str, admin_user_id: str):
     if not env.sendgrid.api_key:
         return password_reset_link
 
-    html_template = email_service.read_email_template("./templates/send_email.html")
-    html_content = html_template.format(
-        username_placeholder=admin_user.username,
-        action_placeholder="requested a password reset for you in their workspace",
-        workspace_placeholder="",
-        call_to_action=f"""<p>Click the link below to reset your password:</p><br><a href="{password_reset_link}">Reset Password</a>""",
-    )
-
-    if not env.sendgrid.from_address:
-        raise ValueError("Sendgrid requires a sender email address to work.")
-
-    await email_service.send_email(
-        from_email=env.sendgrid.from_address,
+    await emailing.send_email(
         to_email=user.email,
         subject=f"{admin_user.username} requested a password reset for you in their workspace",
-        html_content=html_content,
+        username=admin_user.username,
+        action="requested a password reset for you in their workspace",
+        workspace="",
+        call_to_action=f"""<p>Click the link below to reset your password:</p><br><a href="{password_reset_link}">Reset Password</a>""",
     )
