+++ api/oss/tests/pytest/unit/events/test_events_utils.py
@@ -772,9 +772,15 @@ async def test_l1_quota_soft_check_drops_when_over_quota():
         ),
         patch("oss.src.core.events.utils.is_ee", return_value=True),
         patch(
-            "ee.src.utils.entitlements.check_entitlements", new=_fake_check, create=True
+            "ee.src.core.access.entitlements.service.check_entitlements",
+            new=_fake_check,
+            create=True,
         ),
-        patch("ee.src.utils.entitlements.scope_from", new=lambda **kw: kw, create=True),
+        patch(
+            "ee.src.core.access.entitlements.service.scope_from",
+            new=lambda **kw: kw,
+            create=True,
+        ),
     ):
         await publish_trace_fetched(
             request=_make_request(),
@@ -799,9 +805,15 @@ async def test_l1_quota_soft_check_allows_under_quota():
         ),
         patch("oss.src.core.events.utils.is_ee", return_value=True),
         patch(
-            "ee.src.utils.entitlements.check_entitlements", new=_fake_check, create=True
+            "ee.src.core.access.entitlements.service.check_entitlements",
+            new=_fake_check,
+            create=True,
         ),
-        patch("ee.src.utils.entitlements.scope_from", new=lambda **kw: kw, create=True),
+        patch(
+            "ee.src.core.access.entitlements.service.scope_from",
+            new=lambda **kw: kw,
+            create=True,
+        ),
     ):
         await publish_trace_fetched(
             request=_make_request(),
@@ -826,9 +838,15 @@ async def test_l1_quota_soft_check_fails_open_on_exception():
         ),
         patch("oss.src.core.events.utils.is_ee", return_value=True),
         patch(
-            "ee.src.utils.entitlements.check_entitlements", new=_fake_check, create=True
+            "ee.src.core.access.entitlements.service.check_entitlements",
+            new=_fake_check,
+            create=True,
         ),
-        patch("ee.src.utils.entitlements.scope_from", new=lambda **kw: kw, create=True),
+        patch(
+            "ee.src.core.access.entitlements.service.scope_from",
+            new=lambda **kw: kw,
+            create=True,
+        ),
     ):
         await publish_trace_fetched(
             request=_make_request(),
@@ -856,9 +874,15 @@ async def test_l1_quota_soft_check_skipped_when_org_unknown():
         ),
         patch("oss.src.core.events.utils.is_ee", return_value=True),
         patch(
-            "ee.src.utils.entitlements.check_entitlements", new=_fake_check, create=True
+            "ee.src.core.access.entitlements.service.check_entitlements",
+            new=_fake_check,
+            create=True,
         ),
-        patch("ee.src.utils.entitlements.scope_from", new=lambda **kw: kw, create=True),
+        patch(
+            "ee.src.core.access.entitlements.service.scope_from",
+            new=lambda **kw: kw,
+            create=True,
+        ),
     ):
         await publish_trace_fetched(
             request=request,
@@ -886,9 +910,15 @@ async def test_l1_quota_soft_check_skipped_on_oss():
         ),
         patch("oss.src.core.events.utils.is_ee", return_value=False),
         patch(
-            "ee.src.utils.entitlements.check_entitlements", new=_fake_check, create=True
+            "ee.src.core.access.entitlements.service.check_entitlements",
+            new=_fake_check,
+            create=True,
         ),
-        patch("ee.src.utils.entitlements.scope_from", new=lambda **kw: kw, create=True),
+        patch(
+            "ee.src.core.access.entitlements.service.scope_from",
+            new=lambda **kw: kw,
+            create=True,
+        ),
     ):
         await publish_trace_fetched(
             request=_make_request(),
